Added LCD display functionality

+++ src/Si7021.h
@@ -52,8 +52,8 @@ void si7021_init_i2c_temp_read(void);
  *               si7021_init_i2c_temp_read once the I2C transaction is complete.
  *               It takes the 16 bit unsigned int raw output from the si7021 and
  *               converts it into milli-degrees C.
- * @return        [Temperature in degrees C (floating point)]
+ * @return        [Temperature in degrees C (uint32_t)]
  */
-float si7021_return_last_temp(void);
+uint32_t si7021_return_last_temp(void);
 
 #endif /* SRC_SI7021_H_ */
